requirements.txt
Pillow
pyzipper

